Extracted ADREPLDC class to its own file
