+++ ADReplStatus.csproj
@@ -70,6 +70,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ADREPLDC.cs" />
     <Compile Include="ADReplStatusForm.cs">
       <SubType>Form</SubType>
     </Compile>
